Adding toolbar across all activities and adding theme file for UI changes
