+++ app/src/main/AndroidManifest.xml
@@ -10,9 +10,8 @@
         android:name=".utilities.ChatApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/Theme.CHATAPP_THEME">
         <activity android:name=".activities.IndividualChatActivity"/>
         <activity android:name=".activities.SignInActivity">
         <intent-filter>
