+++ app/src/main/res/layout/activity_individual_chat.xml
@@ -1,56 +1,56 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/rl"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
+    android:fitsSystemWindows="true"
     tools:context="com.google.firebase.codelab.friendlychat.activities.IndividualChatActivity">
+
+    <include
+        android:id="@+id/toolbar"
+        layout="@layout/toolbar_main"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/messageRecyclerView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_above="@+id/linearLayout"/>
+        android:layout_above="@+id/linearLayout"
+        android:layout_below="@id/toolbar"
+        android:padding="@dimen/border_margin" />
 
     <LinearLayout
         android:id="@+id/linearLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_gravity="center_vertical"
         android:layout_alignParentBottom="true"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
+        android:layout_gravity="center_vertical"
         android:orientation="horizontal">
 
         <ImageButton
+            android:id="@+id/ibAddTrailor"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@android:drawable/ic_menu_add"
             android:onClick="onAddTrailor"
-            android:id="@+id/ibAddTrailor"/>
+            android:src="@drawable/ic_action_additem" />
 
         <EditText
             android:id="@+id/messageEditText"
-            android:layout_width="0dp"
+            android:layout_width="@dimen/zero_dimen"
             android:layout_height="wrap_content"
             android:layout_gravity="center_vertical"
-            android:layout_weight="1"/>
+            android:layout_weight="1" />
 
         <Button
             android:id="@+id/sendButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-
             android:enabled="false"
-            android:text="@string/send"/>
-
-
-
+            android:text="@string/send" />
     </LinearLayout>
 
 
@@ -60,45 +60,43 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"/>
+        android:layout_centerVertical="true" />
 
     <LinearLayout
         android:id="@+id/llFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
 
-        android:orientation="vertical"
-        android:layout_alignParentTop="true"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:layout_below="@+id/linearLayout"
 
-        android:layout_below="@+id/linearLayout">
+        android:orientation="vertical">
+
         <FrameLayout
             android:id="@+id/flMovieFragment"
             android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-           /> </LinearLayout>
+            android:layout_height="@dimen/zero_dimen"
+            android:layout_weight="1" />
+
+    </LinearLayout>
 
     <LinearLayout
         android:id="@+id/llFragmentPreview"
-        android:layout_width="150dp"
-        android:layout_height="150dp"
-        android:orientation="vertical"
-        android:gravity="center"
+        android:layout_width="@dimen/standard_trailers_size"
+        android:layout_height="@dimen/standard_trailers_size"
         android:layout_above="@+id/linearLayout"
-        android:layout_centerHorizontal="true">
+        android:layout_centerHorizontal="true"
+        android:gravity="center"
+        android:orientation="vertical">
 
+        <FrameLayout
+            android:id="@+id/flMovieFragmentPreview"
+            android:layout_width="@dimen/standard_movie_preview_size"
+            android:layout_height="@dimen/standard_movie_preview_size"
+            android:layout_weight="1"></FrameLayout>
 
-
-    <FrameLayout
-        android:id="@+id/flMovieFragmentPreview"
-        android:layout_width="200dp"
-        android:layout_height="200dp"
-        android:layout_weight="1"
-        >
-    </FrameLayout>
-
-</LinearLayout>
+    </LinearLayout>
 
 </RelativeLayout>
