+++ app/src/main/res/layout/activity_trailer_grid_view.xml
@@ -1,84 +1,50 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    style="@style/general_background"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#f0f0f0">
-
+    android:layout_height="match_parent">
 
     <GridView
         android:id="@+id/gvTrailor"
-        android:layout_below="@+id/llHeader"
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
-        android:layout_margin="5dp"
-        android:columnWidth="100dp"
-        android:drawSelectorOnTop="true"
-        android:gravity="center"
-        android:numColumns="auto_fit"
-        android:stretchMode="columnWidth"
-        android:verticalSpacing="5dp"
-        android:focusable="true"
-        android:clickable="true"/>
+        android:layout_below="@+id/llHeader"
+        style="@style/gridview_style"/>
 
     <RelativeLayout
-
-    android:id="@+id/llHeader"
-
-    android:layout_width="match_parent"
-
-    android:layout_height="wrap_content"
-        android:gravity="center"
-        android:background="#FC9"
-
+        android:id="@+id/llHeader"
+        style="@style/header_background"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
         android:layout_alignParentTop="true"
         android:layout_alignStart="@+id/gvTrailor">
 
-
         <ImageButton
-            android:layout_width="10dp"
-            android:layout_height="10dp"
-            android:backgroundTint="#FC9"
-            android:contentDescription="cancel"
-            android:tint="@color/colorPrimaryDark"
-            app:srcCompat="@android:drawable/ic_menu_close_clear_cancel"
             android:id="@+id/ibCancel"
-            android:onClick="CancelTrailers"
-            android:layout_marginRight="10dp"
-            android:layout_centerVertical="true"
-            android:layout_alignParentEnd="true"
-          />
+            android:layout_width="@dimen/border_margin"
+            android:layout_height="@dimen/border_margin"
+            style="@style/ib_trailer1"
+            app:srcCompat="@android:drawable/ic_menu_close_clear_cancel" />
 
         <TextView
-
-        android:layout_width="match_parent"
-
-        android:layout_height="30dp"
-        android:textAlignment="viewStart"
-
-        android:text="@string/popular_movie_trailer"
-        android:textColor="@color/colorPrimaryDark"
-
-        android:textSize="20sp"
-
             android:id="@+id/tvHeader"
-            android:layout_alignParentTop="true"
+            style="@style/tv_trailer"
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/tv_trailer_ht"
             android:layout_alignParentEnd="true"
-            android:layout_alignParentStart="true" />
+            android:layout_alignParentStart="true"
+            android:layout_alignParentTop="true"
+            android:text="@string/popular_movie_trailer" />
 
         <ImageButton
-
-            android:tint="@color/colorPrimaryDark"
-            android:layout_width="10dp"
-            android:layout_height="10dp"
-            app:srcCompat="@android:drawable/arrow_up_float"
             android:id="@+id/imageButton3"
-            android:background="@android:color/transparent"
-            android:layout_marginTop="10dp"
-            android:layout_alignParentTop="true"
-            android:contentDescription="Expand"
+            style="@style/ib_trailer2"
+            android:layout_width="@dimen/border_margin"
+            android:layout_height="@dimen/border_margin"
             android:layout_toStartOf="@+id/ibCancel"
-            android:layout_marginEnd="11dp"
-            android:onClick="ExpandTrailerView"/>
+            android:contentDescription="@string/expand"
+            android:onClick="ExpandTrailerView"
+            app:srcCompat="@android:drawable/arrow_up_float" />
 
     </RelativeLayout>
