+++ app/src/main/res/layout/item_chat.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools"
     android:padding="@dimen/border_margin">
 
     <de.hdodenhof.circleimageview.CircleImageView
@@ -30,15 +30,15 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_weight="3"
-            tools:text="Firstname Lastname" />
+            tools:text="@string/firstlastname" />
 
         <TextView
             android:id="@+id/tvTimestampChat"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:gravity="end"
             android:layout_weight="1"
-            tools:text="Timestamp" />
+            android:gravity="end"
+            tools:text="@string/timestamp" />
     </LinearLayout>
 
     <TextView
@@ -47,9 +47,9 @@
         android:layout_height="wrap_content"
         android:layout_alignBottom="@id/ivProfileImageChat"
         android:layout_alignParentRight="true"
-        android:scrollHorizontally="true"
-        android:ellipsize="end"
         android:layout_below="@id/llNameTime"
         android:layout_toRightOf="@id/ivProfileImageChat"
-        tools:text="You: Last chat comment by you" />
+        android:ellipsize="end"
+        android:scrollHorizontally="true"
+        tools:text="@string/lastcomment" />
 </RelativeLayout>
