+++ app/src/main/res/layout/item_message_movie.xml
@@ -8,41 +8,39 @@
 
     <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/messengerImageView"
-        android:layout_marginRight="@dimen/activity_horizontal_margin"
         android:layout_width="@dimen/standard_chat_image_size"
         android:layout_height="@dimen/standard_chat_image_size"
+        android:layout_marginRight="@dimen/activity_horizontal_margin"
         android:src="@drawable/ic_account_circle_black_36dp" />
 
     <ImageView
         android:id="@+id/ivTrailerImage"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:minHeight="200dp"
         android:layout_toRightOf="@id/messengerImageView"
         android:contentDescription="@string/trailerposter"
+        android:minHeight="@dimen/standard_movie_preview_size"
         tools:src="@mipmap/ic_launcher" />
 
     <ImageView
         android:id="@+id/ivOverlay"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_alignTop="@id/ivTrailerImage"
         android:layout_alignBottom="@id/ivTrailerImage"
+        android:layout_alignTop="@id/ivTrailerImage"
         android:layout_gravity="center"
+        android:layout_toRightOf="@id/messengerImageView"
         android:contentDescription="@string/overlay_image"
         android:src="@android:drawable/ic_media_play"
-        android:tint="@color/colorPrimaryDark"
-        android:layout_toRightOf="@id/messengerImageView"
-        />
+        android:tint="@color/colorPrimaryDark" />
 
 
     <TextView
         android:id="@+id/messengerTextView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_weight="0"
-        android:layout_toRightOf="@id/messengerImageView"
         android:layout_below="@id/ivTrailerImage"
+        android:layout_toRightOf="@id/messengerImageView"
         android:textAppearance="?android:attr/textAppearanceSmall" />
 
 </RelativeLayout>
