+++ app/src/main/res/layout/item_message_text.xml
@@ -3,8 +3,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:orientation="vertical"
     android:layout_marginTop="@dimen/activity_vertical_margin"
+    android:orientation="vertical"
     tools:targetApi="ice_cream_sandwich">
 
     <LinearLayout
