+++ app/src/main/res/menu/main_menu.xml
@@ -4,7 +4,7 @@
     <item
         android:id="@+id/sign_out_menu"
         android:title="@string/sign_out"
-        app:showAsAction="never"/>
+        app:showAsAction="always"/>
     <item
         android:id="@+id/fresh_config_menu"
         android:title="@string/fresh_config"
@@ -12,7 +12,7 @@
     <item
         android:id="@+id/invite_menu"
         android:title="@string/invite_menu_label"
-        app:showAsAction="never"/>
+        app:showAsAction="always"/>
     <item
         android:id="@+id/crash_menu"
         android:title="@string/cause_crash"/>
