+++ app/src/main/res/values/attrs.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <!-- This is declaring the XML attributes that can be set within a theme. -->
+    <!-- Check the "res/values/themes.xml" file for examples of these attributes being set.  -->
+    <attr name="mainBackground" format="reference" />
+    <attr name="headerBackground" format="reference" />
+    <attr name="toolbarBackground" format="reference" />
+
+    <!-- <attr name="textSubheader" format="reference" />
+    <attr name="textLarge" format="reference" />
+    <attr name="textRegular" format="reference" /> -->
+    <attr name="tvToolbar" format="reference" />
+    <attr name="tvSignin" format="reference" />
+    <attr name="ibCreate" format="reference" />
+    <attr name="ibBack" format="reference" />
+    <attr name="tvContacts" format="reference" />
+    <attr name="ibTrailer1" format="reference" />
+    <attr name="ibTrailer2" format="reference" />
+    <attr name="tvTrailer" format="reference" />
+    <attr name="tvPoster" format="reference" />
+    <attr name="gridViewStyling" format="reference"/>
+    <attr name="ivOverlay" format="reference"/>
+    <attr name="ibCancelTrailer" format="reference"/>
+
+</resources>
