+++ app/src/main/res/values/colors.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#FF9800</color>
-    <color name="colorPrimaryDark">#E65100</color>
-    <color name="colorAccent">#2E7D32</color>
-    <color name="colorTitle">#ffffff</color>
-    <color name="background_white">#FFFFFF</color>
+    <color name="colorPrimary">#39c6b8</color>
+    <color name="colorPrimaryDark">#18a597</color>
+    <color name="colorAccent">#40e0d0</color>
+
+    <color name="white">#ffffff</color>
 
 </resources>
