+++ app/src/main/res/values/dimens.xml
@@ -3,7 +3,18 @@
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
 
+    <dimen name="zero_dimen">0dp</dimen>
+    <dimen name="small_spacing">5dp</dimen>
+    <dimen name="border_margin">10dp</dimen>
+    <dimen name="tv_trailer_ht">30dp</dimen>
+    <dimen name="overlay_top_margin">33dp</dimen>
+    <dimen name="rv_top_margin">60dp</dimen>
+    <dimen name="tv_signin_margin">80dp</dimen>
+
     <dimen name="standard_chat_image_size">36dp</dimen>
     <dimen name="standard_contact_image_size">50dp</dimen>
-    <dimen name="border_margin">10dp</dimen>
+    <dimen name="standard_poster_image_size">100dp</dimen>
+    <dimen name="standard_trailers_size">150dp</dimen>
+    <dimen name="standard_movie_preview_size">200dp</dimen>
+
 </resources>
