+++ build.gradle
@@ -8,7 +8,6 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:2.2.0'
         classpath 'com.google.gms:google-services:3.0.0'
-
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
